feat: allow for manually adding a user to a journey for testing
